V2.1 [1]app模块：    修改music列表与详情    增加category列表接口 [2]console模块：    修改music新增与更新接口，增加参数typeId    增加category的列表、详情、新增、修改、删除接口 [3]通过Spring三级缓存实现循环依赖    music新增与更新接口，增加校验typeId    优化分类删除逻辑

+++ module/src/main/java/top/yuhanpeng/musiccard/module/service/CategoryService.java
@@ -4,6 +4,7 @@ import jakarta.annotation.Resource;
 import org.springframework.stereotype.Service;
 import top.yuhanpeng.musiccard.module.entity.Category;
 import top.yuhanpeng.musiccard.module.mapper.CategoryMapper;
+import top.yuhanpeng.musiccard.module.mapper.MusicMapper;
 
 import java.util.List;
 
@@ -16,6 +17,8 @@ import java.util.List;
 public class CategoryService {
     @Resource
     private CategoryMapper categoryMapper;
+    @Resource
+    private MusicMapper musicMapper;
 
     public Category getById(Long id) throws Exception {
         if (id == null) {
@@ -99,9 +102,13 @@ public class CategoryService {
         return res;
     }
 
-    public Integer delete(Long id) {
+    public Integer delete(Long id) throws Exception {
         if (id == null) {
             throw new RuntimeException("id cannot be null!");
+        }
+        Long countMusicNum = musicMapper.getByTypeId(id);
+        if (countMusicNum != 0) {
+            throw new RuntimeException("there is still music,so this category cannot be deleted");
         }
         int timeStamp = (int) (System.currentTimeMillis() / 1000);
         return categoryMapper.delete(timeStamp, id);
